Added JWT authentication

+++ backend/controller/goalController.js
@@ -1,11 +1,12 @@
 const asyncHandler = require('express-async-handler');
-const Goal = require('../model/goalModel');
+const Goal = require('../models/goalModel');
+const User = require('../models/userModel');
 
 // @access private
 // @route GET /api/goals
 // @desc get goals
 const getGoal = asyncHandler(async(req, res) => {
-    const goals = await Goal.find()
+    const goals = await Goal.find({ user: req.user.id });
     res.status(200).json(goals)
 })
 // @access private
@@ -17,7 +18,8 @@ const setGoal = asyncHandler(async(req, res) => {
         throw new Error('Please add a text field to the request')
     }
     const goal = await Goal.create({
-        text: req.body.text
+        text: req.body.text,
+        user: req.user.id
     })
     res.status(200).json(goal)
 })
@@ -30,6 +32,16 @@ const updateGoal = asyncHandler(async(req, res) => {
     if(!goal) {
         res.status(400)
         throw new Error('Goal not found')
+    }
+    const user = await User.findById(req.user.id)
+
+    if(!user){
+        res.status(401)
+        throw new Error('User not found')
+    }
+
+    if(goal.user.toString() !== user.id) {
+        throw new Error('User not authorized')
     }
 
     const updatedGoal = await Goal.findByIdAndUpdate(req.params.id, req.body, {
@@ -47,6 +59,16 @@ const deleteGoal = asyncHandler(async(req, res) => {
     if(!goal) {
         res.status(400)
         throw new Error('Goal not found')
+    }
+    const user = await User.findById(req.user.id)
+
+    if(!user){
+        res.status(401)
+        throw new Error('User not found')
+    }
+
+    if(goal.user.toString() !== user.id) {
+        throw new Error('User not authorized')
     }
     await Goal.findByIdAndRemove(req.params.id)
